requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.1
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gdown==4.7.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.23.1
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==8.2.0
jupyter-core==5.3.0
kiwisolver==1.4.4
lightning-utilities==0.8.0
lit==16.0.5.post0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==2.0.3
pytorchvideo==0.1.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
segment-anything==1.0
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
spicy==0.16.0
stack-data==0.6.2
submitit==1.4.5
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.1
triton==2.0.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3
wcwidth==0.2.6
Werkzeug==2.3.6
yacs==0.1.8
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
